Add files via upload
This was an end-of-course project for my INF 120 Elementary Programming course at Northern Kentucky University. It had to be a project that solved a "real-world problem" that met a requirement for a certain number of lines of code. This was from April 2021. I have not written much in Python since, but it is the language I started with first.